+++ R/ggbetweenstats.R
@@ -28,6 +28,10 @@
 #' @param outlier.tagging Decides whether outliers should be tagged (Default:
 #'   `FALSE`).
 #' @param outlier.label Label to put on the outliers that have been tagged.
+#' @param outlier.coef Coefficient for outlier detection using Tukey's method.
+#'   With Tukey’s method, outliers are below (1st Quartile) or above (3rd
+#'   Quartile) `outlier.coef` times the Inter-Quartile Range (IQR) (Default:
+#'   `1.5`).
 #' @param mean.plotting Decides whether mean is to be highlighted and its value
 #'   to be displayed (Default: `TRUE`).
 #' @param mean.color Color for the data point corresponding to mean.
@@ -115,7 +119,9 @@ utils::globalVariables(
     "chi_subtitle",
     "proptest_subtitle",
     "LL",
-    "UL"
+    "UL",
+    "outlier",
+    "outlier.label"
   )
 )
 
@@ -135,6 +141,7 @@ ggbetweenstats <- function(data = NULL,
                            outlier.tagging = NULL,
                            outlier.label = NULL,
                            outlier.color = "black",
+                           outlier.coef = 1.5,
                            mean.plotting = TRUE,
                            mean.color = "darkred",
                            messages = TRUE) {
@@ -146,11 +153,9 @@ ggbetweenstats <- function(data = NULL,
   # if dataframe is provided
   if (!is.null(data)) {
     # preparing labels from given dataframe
-    lab.df <- colnames(dplyr::select(
-      .data = data,
-      !!rlang::enquo(x),
-      !!rlang::enquo(y)
-    ))
+    lab.df <- colnames(dplyr::select(.data = data,
+                                     !!rlang::enquo(x),
+                                     !!rlang::enquo(y)))
     # if xlab is not provided, use the variable x name
     if (is.null(xlab)) {
       xlab <- lab.df[1]
@@ -182,18 +187,14 @@ ggbetweenstats <- function(data = NULL,
     if (!is.null(outlier.label)) {
       # if vectors are provided and outlier label vector is present
       data <-
-        base::cbind.data.frame(
-          x = x,
-          y = y,
-          outlier.label = outlier.label
-        )
+        base::cbind.data.frame(x = x,
+                               y = y,
+                               outlier.label = outlier.label)
     } else {
       # if outlier label vector is absent
       data <-
-        base::cbind.data.frame(
-          x = x,
-          y = y
-        )
+        base::cbind.data.frame(x = x,
+                               y = y)
     }
   }
   # x needs to be a factor for group or condition comparison
@@ -208,7 +209,7 @@ ggbetweenstats <- function(data = NULL,
       dplyr::mutate_at(
         .tbl = .,
         .vars = "x",
-        .funs = ~base::droplevels(x = .)
+        .funs = ~ base::droplevels(x = .)
       )
   } else if (!is.factor(data$x)) {
     # convert to factor
@@ -219,7 +220,7 @@ ggbetweenstats <- function(data = NULL,
       dplyr::mutate_at(
         .tbl = .,
         .vars = "x",
-        .funs = ~base::droplevels(x = .)
+        .funs = ~ base::droplevels(x = .)
       )
     # display message
     base::message(cat(
@@ -324,11 +325,9 @@ ggbetweenstats <- function(data = NULL,
         # partial omega-squared is the biased estimate of effect size for parametric ANOVA
         # computing confidence interval for omega-squared
         aov_effsize_ci <-
-          userfriendlyscience::confIntOmegaSq(
-            var1 = data$x,
-            var2 = data$y,
-            conf.level = 0.95
-          )
+          userfriendlyscience::confIntOmegaSq(var1 = data$x,
+                                              var2 = data$y,
+                                              conf.level = 0.95)
 
         # aov_stat input represents the anova object summary derived from car library
         rsubtitle_omega <-
@@ -349,7 +348,7 @@ ggbetweenstats <- function(data = NULL,
                   " = ",
                   pvalue,
                   ", ",
-                  italic(omega)^2,
+                  italic(omega) ^ 2,
                   " = ",
                   effsize,
                   ", 95% CI [",
@@ -374,10 +373,8 @@ ggbetweenstats <- function(data = NULL,
         # adding the subtitle to the plot
         plot <-
           plot +
-          ggplot2::labs(subtitle = rsubtitle_omega(
-            aov_stat = aov_stat,
-            aov_effsize_ci = aov_effsize_ci
-          ))
+          ggplot2::labs(subtitle = rsubtitle_omega(aov_stat = aov_stat,
+                                                   aov_effsize_ci = aov_effsize_ci))
       } else if (effsize.type == "biased") {
         # partial eta-squared is the biased estimate of effect size for parametric ANOVA
         aov_effsize <-
@@ -390,10 +387,8 @@ ggbetweenstats <- function(data = NULL,
 
         # aov_stat2 object is *only* to compute partial eta-squared since there is no straightforward to get partial
         # eta-squared for Welch's ANOVA and its confidence interval
-        aov_stat2 <- summary(stats::aov(
-          formula = y ~ x,
-          data = data
-        ))
+        aov_stat2 <- summary(stats::aov(formula = y ~ x,
+                                        data = data))
 
         # getting confidence interval for partial eta-squared
         aov_effsize_ci <- apaTables::get.ci.partial.eta.squared(
@@ -409,8 +404,8 @@ ggbetweenstats <- function(data = NULL,
         # aov_stat input represents the anova object summary derived from car library
         rsubtitle_peta <-
           function(aov_stat,
-                             aov_effsize,
-                             aov_effsize_ci) {
+                   aov_effsize,
+                   aov_effsize_ci) {
             # extracting the elements of the statistical object
             base::substitute(
               expr =
@@ -427,7 +422,7 @@ ggbetweenstats <- function(data = NULL,
                   " = ",
                   pvalue,
                   ", p",
-                  italic(eta)^2,
+                  italic(eta) ^ 2,
                   " = ",
                   effsize,
                   ", 95% CI [",
@@ -472,11 +467,9 @@ ggbetweenstats <- function(data = NULL,
     } else if (type == "nonparametric" || type == "np") {
       ############################ Kruskal-Wallis (nonparametric ANOVA) #################################################
       # setting up the anova model and getting its summary
-      kw_stat <- stats::kruskal.test(
-        formula = y ~ x,
-        data = data,
-        na.action = na.omit
-      )
+      kw_stat <- stats::kruskal.test(formula = y ~ x,
+                                     data = data,
+                                     na.action = na.omit)
 
       # aov_stat input represents the anova object summary derived from car library
       rsubtitle_kw <- function(kw_stat) {
@@ -485,7 +478,7 @@ ggbetweenstats <- function(data = NULL,
           expr =
             paste(
               "Kruskal-Wallis: ",
-              italic(chi)^2,
+              italic(chi) ^ 2,
               "(",
               df,
               ") = ",
@@ -499,11 +492,9 @@ ggbetweenstats <- function(data = NULL,
             estimate = ggstatsplot::specify_decimal_p(x = kw_stat$statistic[[1]], k),
             df = kw_stat$parameter[[1]],
             # degrees of freedom are always integer
-            pvalue = ggstatsplot::specify_decimal_p(
-              x = kw_stat$p.value[[1]],
-              k,
-              p.value = TRUE
-            )
+            pvalue = ggstatsplot::specify_decimal_p(x = kw_stat$p.value[[1]],
+                                                    k,
+                                                    p.value = TRUE)
           )
         )
       }
@@ -548,11 +539,9 @@ ggbetweenstats <- function(data = NULL,
             df1 = robust_aov_stat$df1[[1]],
             # degrees of freedom are always integer
             df2 = ggstatsplot::specify_decimal_p(x = robust_aov_stat$df2[[1]], k),
-            pvalue = ggstatsplot::specify_decimal_p(
-              x = robust_aov_stat$p.value[[1]],
-              k,
-              p.value = TRUE
-            ),
+            pvalue = ggstatsplot::specify_decimal_p(x = robust_aov_stat$p.value[[1]],
+                                                    k,
+                                                    p.value = TRUE),
             effsize = ggstatsplot::specify_decimal_p(x = robust_aov_stat$effsize[[1]], k)
           )
         )
@@ -640,10 +629,8 @@ ggbetweenstats <- function(data = NULL,
         # adding subtitle to the plot
         plot <-
           plot +
-          ggplot2::labs(subtitle = rsubtitle_g(
-            t_stat = t_stat,
-            t_effsize = t_effsize
-          ))
+          ggplot2::labs(subtitle = rsubtitle_g(t_stat = t_stat,
+                                               t_effsize = t_effsize))
       } else if (effsize.type == "biased") {
         # t_stat input represents the t-test object summary derived from stats library
         rsubtitle_d <- function(t_stat, t_effsize) {
@@ -694,10 +681,8 @@ ggbetweenstats <- function(data = NULL,
         # adding subtitle to the plot
         plot <-
           plot +
-          ggplot2::labs(subtitle = rsubtitle_d(
-            t_stat = t_stat,
-            t_effsize = t_effsize
-          ))
+          ggplot2::labs(subtitle = rsubtitle_d(t_stat = t_stat,
+                                               t_effsize = t_effsize))
       }
 
       # displaying the details of the test that was run
@@ -773,10 +758,8 @@ ggbetweenstats <- function(data = NULL,
       # adding subtitle to the plot
       plot <-
         plot +
-        ggplot2::labs(subtitle = rsubtitle_mann(
-          mann_stat = mann_stat,
-          z_stat = z_stat
-        ))
+        ggplot2::labs(subtitle = rsubtitle_mann(mann_stat = mann_stat,
+                                                z_stat = z_stat))
     } else if (type == "robust" || type == "r") {
       ######################################### robust t-test ############################################################
 
@@ -809,11 +792,9 @@ ggbetweenstats <- function(data = NULL,
             env = base::list(
               estimate = ggstatsplot::specify_decimal_p(x = t_robust_stat$test[[1]], k),
               df = ggstatsplot::specify_decimal_p(x = t_robust_stat$df[[1]], k),
-              pvalue = ggstatsplot::specify_decimal_p(
-                x = t_robust_stat$p.value[[1]],
-                k,
-                p.value = TRUE
-              ),
+              pvalue = ggstatsplot::specify_decimal_p(x = t_robust_stat$p.value[[1]],
+                                                      k,
+                                                      p.value = TRUE),
               effsize = ggstatsplot::specify_decimal_p(x = t_robust_effsize$effsize[[1]], k),
               LL = ggstatsplot::specify_decimal_p(x = t_robust_effsize$CI[[1]][[1]], k),
               UL = ggstatsplot::specify_decimal_p(x = t_robust_effsize$CI[[2]][[1]], k)
@@ -823,29 +804,22 @@ ggbetweenstats <- function(data = NULL,
 
       # setting up the independent samples t-tests on robust location measures (without bootstraps)
       t_robust_stat <-
-        WRS2::yuen(
-          formula = y ~ x,
-          data = data
-        )
+        WRS2::yuen(formula = y ~ x,
+                   data = data)
       # computing effect sizes
       t_robust_effsize <-
-        WRS2::yuen.effect.ci(
-          formula = y ~ x,
-          data = data
-        )
+        WRS2::yuen.effect.ci(formula = y ~ x,
+                             data = data)
 
       # adding the label to the plot
       plot <-
         plot +
-        ggplot2::labs(subtitle = rsubtitle_rob(
-          t_robust_stat = t_robust_stat,
-          t_robust_effsize = t_robust_effsize
-        ))
+        ggplot2::labs(subtitle = rsubtitle_rob(t_robust_stat = t_robust_stat,
+                                               t_robust_effsize = t_robust_effsize))
     }
   }
 
   ########################################### outlier tagging #########################################################
-
 
   # if outlier.tagging is set to TRUE, first figure out what labels need to be attached to the outlier
   if (isTRUE(outlier.tagging)) {
@@ -853,11 +827,9 @@ ggbetweenstats <- function(data = NULL,
     if (missing(outlier.label)) {
       # if outlier label is not provided, outlier labels will just be values of the y vector
       data_df <-
-        base::cbind.data.frame(
-          x = data$x,
-          y = data$y,
-          outlier.label = data$y
-        )
+        base::cbind.data.frame(x = data$x,
+                               y = data$y,
+                               outlier.label = data$y)
     } else {
       # if the outlier tag has been provided, just use the dataframe already created
       data_df <-
@@ -867,32 +839,15 @@ ggbetweenstats <- function(data = NULL,
           outlier.label = data$outlier.label
         )
     }
-    ## finding the outliers in the dataframe using Tukey's interquartile range rule
-
-    # defining function to detect outliers
-    check_outlier <- function(v, coef = 1.5) {
-      # compute the quantiles
-      quantiles <- stats::quantile(
-        x = v,
-        probs = c(0.25, 0.75)
-      )
-      # compute the interquartile range
-      IQR <- quantiles[2] - quantiles[1]
-      # check for outlier and output a logical
-      res <-
-        ((v < (quantiles[1] - coef * IQR)) |
-          (v > (quantiles[2] + coef * IQR)))
-      # return the result
-      return(res)
-    }
 
     # finding and tagging the outliers
-    data_df <- data_df %>%
+    data_df %<>%
       dplyr::group_by(.data = ., x) %>%
       dplyr::mutate(
         .data = .,
         outlier = base::ifelse(
-          test = check_outlier(y),
+          test = check_outlier(var = y,
+                               coef = outlier.coef),
           yes = outlier.label,
           no = NA
         )
@@ -904,14 +859,27 @@ ggbetweenstats <- function(data = NULL,
     if (is.character(data_df$outlier.label)) {
       data_df$outlier.label <- as.factor(data_df$outlier.label)
     }
+
     # if outlier labels are words or other types of characters, you want these characters to be diaplyed and not the values
     if (is.factor(data_df$outlier.label)) {
       data_df$outlier.label <- as.character(data_df$outlier.label)
+
+      # convert outlier.label to NAs when outlier is also NA
+      data_df %<>%
+        dplyr::mutate(
+          .data = .,
+          outlier = base::ifelse(
+            test = !is.na(outlier),
+            yes = outlier.label,
+            no = NA
+          )
+        )
+      print(data_df)
       # applying the labels to tagged outliers with ggrepel
       plot <-
         plot +
         ggrepel::geom_label_repel(
-          mapping = ggplot2::aes(label = data_df$outlier.label),
+          mapping = ggplot2::aes(label = data_df$outlier),
           fontface = "bold",
           color = "black",
           max.iter = 3e2,
@@ -955,16 +923,14 @@ ggbetweenstats <- function(data = NULL,
     # create a dataframe with means
     mean_dat <- data %>%
       dplyr::group_by(.data = ., x) %>% # group by the independent variable
-      dplyr::mutate_all(
-        .tbl = .,
-        .funs = mean,
-        na.rm = TRUE
-      ) %>% # dependent variable mean for each level of grouping variable
+      dplyr::mutate_all(.tbl = .,
+                        .funs = mean,
+                        na.rm = TRUE) %>% # dependent variable mean for each level of grouping variable
       dplyr::distinct(.data = .) %>% # removed duplicated rows
       dplyr::mutate_if(
         .tbl = .,
         .predicate = is.numeric,
-        .funs = ~as.numeric(as.character(
+        .funs = ~ as.numeric(as.character(
           ggstatsplot::specify_decimal_p(x = ., k = k)
         )) # format the values for printing
       ) %>%
@@ -1005,19 +971,15 @@ ggbetweenstats <- function(data = NULL,
           ": p-value = "
         ),
         crayon::yellow(
-          ggstatsplot::specify_decimal_p(
-            x = ad_norm$p.value[[1]],
-            k,
-            p.value = TRUE
-          )
+          ggstatsplot::specify_decimal_p(x = ad_norm$p.value[[1]],
+                                         k,
+                                         p.value = TRUE)
         )
       ))
     }
     # homogeneity of variance
-    bartlett <- stats::bartlett.test(
-      formula = y ~ x,
-      data = data
-    )
+    bartlett <- stats::bartlett.test(formula = y ~ x,
+                                     data = data)
     # display homogeneity of variances test result as a message
     base::message(cat(
       crayon::green("Note: "),
@@ -1028,11 +990,9 @@ ggbetweenstats <- function(data = NULL,
         ": p-value = "
       ),
       crayon::yellow(
-        ggstatsplot::specify_decimal_p(
-          x = bartlett$p.value[[1]],
-          k,
-          p.value = TRUE
-        )
+        ggstatsplot::specify_decimal_p(x = bartlett$p.value[[1]],
+                                       k,
+                                       p.value = TRUE)
       )
     ))
   }